+++ packages/build-field-types/__fixtures__/basic-field-type/src/index.js
@@ -1,0 +1,7 @@
+import { importView } from '@keystone-alpha/build-field-types';
+
+export let MyCoolFieldType = {
+  views: {
+    Field: importView('./views/Field'),
+  },
+};