+++ packages/build-field-types/src/types.js
@@ -1,0 +1,10 @@
+// @flow
+
+/*::
+import {Project} from './project'
+import {Package} from './package'
+import {Entrypoint} from './entrypoint'
+
+*/
+
+export type ItemUnion = Project | Package | Entrypoint;