+++ packages/fields/src/types/Content/blocks/embed.js
@@ -1,0 +1,6 @@
+import { importView } from '@keystone-alpha/build-field-types';
+
+export default {
+  type: 'embed',
+  viewPath: importView('../views/editor/blocks/embed'),
+};