+++ packages/fields/src/types/Url/index.js
@@ -1,0 +1,17 @@
+import { Text, MongoTextInterface, KnexTextInterface } from '../Text/Implementation';
+import { importView } from '@keystone-alpha/build-field-types';
+
+export default {
+  type: 'Url',
+  implementation: Text,
+  views: {
+    Controller: importView('../Text/views/Controller'),
+    Field: importView('./views/Field'),
+    Filter: importView('../Text/views/Filter'),
+    Cell: importView('./views/Cell'),
+  },
+  adapters: {
+    mongoose: MongoTextInterface,
+    knex: KnexTextInterface,
+  },
+};