+++ packages/arch/packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @arch-ui/common
 
+## 0.0.3
+
+- [patch][e75c105c](https://github.com/keystonejs/keystone-5/commit/e75c105c):
+
+  - admin revamp
+
 ## 0.0.2
 
 - Updated dependencies [23c3fee5]: