+++ test-projects/twitter-login/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @keystone-alpha/cypress-project-twitter-login
 
+## 1.0.7
+
+- Updated dependencies [85b74a2c](https://github.com/keystonejs/keystone-5/commit/85b74a2c):
+- Updated dependencies [e75c105c](https://github.com/keystonejs/keystone-5/commit/e75c105c):
+  - @keystone-alpha/admin-ui@3.1.0
+  - @keystone-alpha/keystone@3.1.0
+  - @keystone-alpha/fields@5.0.0
+  - @arch-ui/fields@0.0.4
+  - @arch-ui/input@0.0.4
+
 ## 1.0.6
 
 - [patch][b4dcf44b](https://github.com/keystonejs/keystone-5/commit/b4dcf44b):